+++ Automation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.4.0 (TBD):
 * [Feature #83] Add the `BreedingPod.Progress` signal to track breeding pod completion progress.
+* [Feature] Add a setting to disable game automation conflict prevention.
 
 # v4.3.0 (June 18th, 2026):
 * [Feature] Add actions for Fill Valve and Throttling Valve.